# Avoid allocation of nulls in setNull