### **ONLY LAST COMMIT BELONGS TO THIS DIFF**
**CLICK ON COMMITS CHOOSE LAST COMMIT AND REVIEW THAT ONLY**

When setNull is used to set an index to not null, it should not allocate nulls if they are not already allocated.

Those numbers are for the null-free case.
now SimpleGeneral and SimpleProxyPushBack are faster as well than SimpleOld for null-free case.

**NO NULLS**

```
After
============================================================================
VectorBasic                                                101.38ns    9.86M
VectorResizeOptimized                                       90.19ns   11.09M
SimpleProxyWithResize                                      116.41ns    8.59M
SimpleProxyPushBack                                        310.51ns    3.22M
SimpleGeneral                                              333.37ns    3.00M
SimpleOld                                                  514.16ns    1.94M
============================================================================
```

```
Before
============================================================================
VectorBasic                                                101.06ns    9.89M
VectorResizeOptimized                                       90.48ns   11.05M
SimpleProxyWithResize                                      128.08ns    7.81M
SimpleProxyPushBack                                        868.48ns    1.15M
SimpleGeneral                                              877.75ns    1.14M
SimpleOld                                                  668.36ns    1.50M
============================================================================
```

**WITH NULLS CASE**
we see also speedups here because while array has nulls the outer most vector is null free.

```
After:
============================================================================
VectorBasic                                                264.58ns    3.78M
VectorResizeOptimized                                      213.93ns    4.67M
SimpleProxyWithResize                                      277.15ns    3.61M
SimpleProxyPushBack                                        732.07ns    1.37M
SimpleGeneral                                              771.50ns    1.30M
SimpleOld                                                  637.80ns    1.57M
============================================================================
```

```
Before:
============================================================================
VectorBasic                                                386.95ns    2.58M
VectorResizeOptimized                                      266.95ns    3.75M
SimpleProxyWithResize                                      288.13ns    3.47M
SimpleProxyPushBack                                        925.42ns    1.08M
SimpleGeneral                                                1.01us  988.39K
SimpleOld                                                  767.64ns    1.30M
============================================================================
```